discord-bot/requirements.txt
# Discord Bot Requirements
discord.py>=2.3.2
python-dotenv>=1.0.0

# Data processing (already in main requirements.txt)
pandas>=2.0.0
numpy>=1.24.0

# Plotting
plotly>=5.14.0
kaleido>=0.2.1

# API clients (already in main requirements.txt)
requests>=2.31.0
aiohttp>=3.9.0

# Existing project dependencies (reference parent requirements.txt)
# The bot will use the parent project's src/ code
